Refactoring and minor fix
