+++ src/redux/actions/samples/createSamples.js
@@ -101,21 +101,24 @@ const createSamples = (
     // order is preserved in newSamples
     const sampleIdsByName = { ...alreadyCreatedSampleIds, ...newSampleIdsByName };
 
-    const newSamplesToRedux = newSamples.map(([name, { files }]) => ({
-      ..._.cloneDeep(sampleTemplate),
-      name,
-      type: sampleTechnology,
-      experimentId,
-      uuid: sampleIdsByName[name],
-      createdDate,
-      lastModified: createdDate,
-      options,
-      metadata: experiment?.metadataKeys
-        .reduce((acc, curr) => ({ ...acc, [curr]: METADATA_DEFAULT_VALUE }), {}) || {},
-      files: Object.values(files).reduce(((acc, curr) => (
-        { ...acc, [curr.name]: { upload: { status: UploadStatus.UPLOADING } } }
-      )), {}),
-    }));
+    const newSamplesToRedux = newSamples
+      // Remove repeated samples
+      .filter(([name]) => newSampleIdsByName[name])
+      .map(([name, { files }]) => ({
+        ..._.cloneDeep(sampleTemplate),
+        name,
+        type: sampleTechnology,
+        experimentId,
+        uuid: sampleIdsByName[name],
+        createdDate,
+        lastModified: createdDate,
+        options,
+        metadata: experiment?.metadataKeys
+          .reduce((acc, curr) => ({ ...acc, [curr]: METADATA_DEFAULT_VALUE }), {}) || {},
+        files: Object.values(files).reduce(((acc, curr) => (
+          { ...acc, [curr.name]: { upload: { status: UploadStatus.UPLOADING } } }
+        )), {}),
+      }));
 
     dispatch({
       type: SAMPLES_CREATED,
