--- src/redux/actions/samples/updateSamplesValidating.js
@@ -1,12 +1,0 @@
-import {
-  SAMPLES_VALIDATING_UPDATED,
-} from 'redux/actionTypes/samples';
-
-const updateSamplesValidating = (experimentId, validating) => async (dispatch) => {
-  dispatch({
-    type: SAMPLES_VALIDATING_UPDATED,
-    payload: { experimentId, validating },
-  });
-};
-
-export default updateSamplesValidating;
